chore: abstract choice array animation
